requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.9.0
numpy<2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
PyPDF2>=3.0.0

